refactor(sync): move addSelectorOptions into commands/sync.ts
Pull the per-kind selector flag builder out of lib/hosts/option.ts
(where it was misplaced) and define it locally in commands/sync.ts.
It is specific to agents sync and has no other callers.

lib/hosts/option.ts now exports only addHostOption, matching origin/main.
This also removes hosts/ from the Windows CI scope trigger, sidestepping
a pre-existing exec.test.ts assertion bug (commit 7757c5b7) that fires
only when that path is touched.
